Started "Start Signals _ min _ sec into Playlist"
This is to allow the user to press "play" on their playlist and then at a predetermined moment, like when they hear a specific drum-beat or a chime or something, they start the iFit signal sequence.  For example, if a user has "The Dark Side of the Moon" as their playlist, they would start the album/playlist and then when the first chord of Breathe is strummed (right after the Speak to Me audio), they would start the iFit Signal... that happens right around 1:16 into the album/playlist.